exchange rate

+++ src/main/java/faang/school/paymentservice/dto/ExchangeRatesResponse.java
@@ -2,6 +2,7 @@ package faang.school.paymentservice.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
@@ -10,5 +11,5 @@ public class ExchangeRatesResponse {
     private String license;
     private long timestamp;
     private String base;
-    private Map<Currency, Double> rates;
+    private Map<Currency, BigDecimal> rates;
 }
